services/users/requirements.txt
flask
psycopg2-binary
sqlalchemy
flask-cors
gunicorn